requirements..txt
click
Pillow
matplotlib
pycryptodome
